requirements.txt
python-binance
#TA-Lib
numpy
backtrader
flask
Flask-Cors
